Harden Parquet direct-buffer wrap helpers
  Use Math.toIntExact for the int-narrowing of BytesInput#size()
  in PrefetchedPageReader.decompressToDirectBuffer so a >2GB page
  fails loudly instead of silently wrapping.
  Widen bytesInputFrom/bytesInputFromSlice to package-private and
  add PrefetchedRowGroupBuilderBytesInputTests covering the
  wrap-don't-copy invariant for direct buffers, plus the zero-length
  slice early return. No-copy is verified by mutating the source
  after wrapping and asserting the BytesInput sees the change.

+++ x-pack/plugin/esql-datasource-parquet/src/main/java/org/elasticsearch/xpack/esql/datasource/parquet/PrefetchedPageReader.java
@@ -204,7 +204,7 @@ final class PrefetchedPageReader implements PageReader {
             input = directInput;
         }
         ByteBuffer output = ByteBuffer.allocateDirect(decompressedSize);
-        decompressor.decompress(input, (int) compressed.size(), output, decompressedSize);
+        decompressor.decompress(input, Math.toIntExact(compressed.size()), output, decompressedSize);
         output.flip();
         return BytesInput.from(output);
     }
